+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/user/UserLikeMemberDao.java
@@ -4,11 +4,11 @@ import com.bunnyxt.tdd.model.user.UserLikeMember;
 
 public interface UserLikeMemberDao {
 
-    void addUserLikeMember(Integer added, Long userid, Integer mid);
+    void addUserLikeMember(Integer added, Long userid, Long mid);
 
-    Integer deleteUserLikeMember(Long userid, Integer mid);
+    Integer deleteUserLikeMember(Long userid, Long mid);
 
-    UserLikeMember queryUserLikeMember(Long userid, Integer mid);
+    UserLikeMember queryUserLikeMember(Long userid, Long mid);
 
-    Integer queryUserLikeMemberCount(Integer mid);
+    Integer queryUserLikeMemberCount(Long mid);
 }
