+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/model/Member.java
@@ -4,7 +4,7 @@ public class Member {
 
     private Long id;
     private Integer added;
-    private Integer mid;
+    private Long mid;
     private String sex;
     private String name;
     private String face;
@@ -32,11 +32,11 @@ public class Member {
         this.added = added;
     }
 
-    public Integer getMid() {
+    public Long getMid() {
         return mid;
     }
 
-    public void setMid(Integer mid) {
+    public void setMid(Long mid) {
         this.mid = mid;
     }
