+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/model/user/UserLikeMember.java
@@ -5,7 +5,7 @@ public class UserLikeMember {
     private Long id;
     private Integer added;
     private Long userid;
-    private Integer mid;
+    private Long mid;
 
     public Long getId() {
         return id;
@@ -31,11 +31,11 @@ public class UserLikeMember {
         this.userid = userid;
     }
 
-    public Integer getMid() {
+    public Long getMid() {
         return mid;
     }
 
-    public void setMid(Integer mid) {
+    public void setMid(Long mid) {
         this.mid = mid;
     }
 }
