+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/MemberFollowerRecordService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface MemberFollowerRecordService {
 
-    List<MemberFollowerRecord> queryMemberFollowerRecords(Integer mid, Integer last_record, Integer start_ts, Integer end_ts,
+    List<MemberFollowerRecord> queryMemberFollowerRecords(Long mid, Integer last_record, Integer start_ts, Integer end_ts,
                                                           Integer pn, Integer ps);
 
-    Integer queryMemberFollowerRecordsCount(Integer mid, Integer start_ts, Integer end_ts);
+    Integer queryMemberFollowerRecordsCount(Long mid, Integer start_ts, Integer end_ts);
 }
