+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/user/UserFavoriteMemberService.java
@@ -7,16 +7,16 @@ import java.util.List;
 
 public interface UserFavoriteMemberService {
 
-    TddCommonResponse postUserFavoriteMember(Long userid, Integer mid);
+    TddCommonResponse postUserFavoriteMember(Long userid, Long mid);
 
-    TddCommonResponse deleteUserFavoriteMember(Long userid, Integer mid);
+    TddCommonResponse deleteUserFavoriteMember(Long userid, Long mid);
 
-    UserFavoriteMember queryUserFavoriteMember(Long userid, Integer mid);
+    UserFavoriteMember queryUserFavoriteMember(Long userid, Long mid);
 
-    Integer queryUserFavoriteMemberCount(Integer mid);
+    Integer queryUserFavoriteMemberCount(Long mid);
 
-    List<UserFavoriteMember> queryUserFavoriteMemberUsers(Integer mid, Integer start_ts, Integer end_ts,
+    List<UserFavoriteMember> queryUserFavoriteMemberUsers(Long mid, Integer start_ts, Integer end_ts,
                                                           String order_by, Integer desc, Integer pn, Integer ps);
 
-    Integer queryUserFavoriteMemberUsersCount(Integer mid, Integer start_ts, Integer end_ts);
+    Integer queryUserFavoriteMemberUsersCount(Long mid, Integer start_ts, Integer end_ts);
 }
